# backend/hotel-api-spec.json
{"openapi":"3.0.0","info":{"title":"Hotel Booking API","version":"1.0.0","description":"A comprehensive API for hotel room booking and management system","contact":{"name":"Hotel API Support","email":"support@hotel.com"}},"servers":[{"url":"http://localhost:3001","description":"Development server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"User":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string","format":"email"},"role":{"type":"string","enum":["client","manager"]}}},"Room":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"image":{"type":"string"},"capacity":{"type":"number"},"size":{"type":"string"},"originalPrice":{"type":"number"},"currentPrice":{"type":"number"},"taxes":{"type":"number"},"total":{"type":"number"},"description":{"type":"string"},"amenities":{"type":"array","items":{"type":"string"}},"availability":{"type":"string","enum":["Available","Limited","Unavailable"]},"isActive":{"type":"boolean"},"createdBy":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"Booking":{"type":"object","properties":{"_id":{"type":"string"},"room":{"type":"string"},"user":{"type":"string"},"checkIn":{"type":"string","format":"date"},"checkOut":{"type":"string","format":"date"},"guests":{"type":"number"},"totalAmount":{"type":"number"},"status":{"type":"string","enum":["pending","confirmed","cancelled","completed"]},"specialRequests":{"type":"string"},"guestDetails":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"}}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ApiResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object"},"count":{"type":"number"}}}}},"security":[{"bearerAuth":[]}],"paths":{"/api/auth/register":{"post":{"summary":"Register a new user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["firstName","lastName","email","password"],"properties":{"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"email":{"type":"string","format":"email","example":"john.doe@example.com"},"password":{"type":"string","minLength":6,"example":"password123"},"role":{"type":"string","enum":["client","manager"],"default":"client","example":"client"}}}}}},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"},"example":{"success":true,"message":"User registered successfully","data":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","user":{"id":"507f1f77bcf86cd799439011","firstName":"John","lastName":"Doe","email":"john.doe@example.com","role":"client"}}}}}},"400":{"description":"Bad request - validation error"},"409":{"description":"User already exists"}}}},"/api/auth/login":{"post":{"summary":"Login user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email","example":"john.doe@example.com"},"password":{"type":"string","example":"password123"},"role":{"type":"string","enum":["client","manager"],"example":"client"}}}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"},"example":{"success":true,"message":"Login successful","data":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","user":{"id":"507f1f77bcf86cd799439011","firstName":"John","lastName":"Doe","email":"john.doe@example.com","role":"client"}}}}}},"400":{"description":"Bad request"},"401":{"description":"Invalid credentials"}}}},"/api/auth/me":{"get":{"summary":"Get current user profile","tags":["Authentication"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"User profile retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"},"example":{"success":true,"data":{"user":{"id":"507f1f77bcf86cd799439011","firstName":"John","lastName":"Doe","email":"john.doe@example.com","role":"client"}}}}}},"401":{"description":"Not authorized"}}}},"/api/rooms":{"get":{"summary":"Get all active rooms","tags":["Rooms"],"responses":{"200":{"description":"Rooms retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"},"example":{"success":true,"count":2,"data":[{"_id":"507f1f77bcf86cd799439011","name":"Deluxe Suite","capacity":2,"currentPrice":15000,"availability":"Available"}]}}}},"500":{"description":"Server error"}}},"post":{"summary":"Create a new room (Manager only)","tags":["Rooms"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"responses":{"201":{"description":"Room created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"400":{"description":"Bad request - validation error"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden - Manager role required"}}}},"/api/rooms/{id}":{"get":{"summary":"Get a single room by ID","tags":["Rooms"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"Room ID"}],"responses":{"200":{"description":"Room retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"404":{"description":"Room not found"},"500":{"description":"Server error"}}},"put":{"summary":"Update a room (Manager only)","tags":["Rooms"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"Room ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"responses":{"200":{"description":"Room updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"400":{"description":"Bad request - validation error"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden - Manager role required"},"404":{"description":"Room not found"}}},"delete":{"summary":"Delete a room (Manager only)","tags":["Rooms"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"Room ID"}],"responses":{"200":{"description":"Room deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"401":{"description":"Not authorized"},"403":{"description":"Forbidden - Manager role required"},"404":{"description":"Room not found"},"500":{"description":"Server error"}}}},"/api/rooms/{id}/availability":{"post":{"summary":"Check room availability for given dates","tags":["Rooms"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"Room ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["checkIn","checkOut"],"properties":{"checkIn":{"type":"string","format":"date","example":"2024-01-15"},"checkOut":{"type":"string","format":"date","example":"2024-01-17"}}}}}},"responses":{"200":{"description":"Availability checked successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"},"example":{"success":true,"data":{"isAvailable":true,"conflictingBookings":0}}}}},"400":{"description":"Bad request - invalid dates"},"500":{"description":"Server error"}}}},"/api/rooms/manager/{managerId}":{"get":{"summary":"Get rooms by manager ID","tags":["Rooms"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"managerId","required":true,"schema":{"type":"string"},"description":"Manager ID"}],"responses":{"200":{"description":"Manager rooms retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"401":{"description":"Not authorized"},"500":{"description":"Server error"}}}}},"tags":[]}